Finish the assignment code
Test have been add and ran (with result tested manually). All test
proved that at least for what I can think of, the program behavior
complies with the specification and are correct.

For the `simulate` function and all associated object, valgrind is used
to test for memory leak. Result shows that for all the test input, the
program does no leak any memory.